moved initiatialization of bootstrapping into cloud runner
